refactor: remove utils module
Small refactoring, consolidate methods and functions inside of the
`settings` module.
